requirements.txt
Django==4.1.13
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
model-bakery==1.17.0
pandas==2.1.3
pytest-django==4.7.0
django-model-utils==4.3.1
drf-spectacular==0.26.5
Faker==20.1.0
pytest~=7.2.0